multi: OnionPeerActor for per-peer message support
Introduce OnionPeerActor to handle the sending of onion message to each
peer. The actor is registered with the receptionist pattern to enable
message routing through the actor system. Also adds onion message
feature bits to the protocol, so that the actor is only spawned when the
peer supports onion messages.